pyproject.toml
[build-system]
requires = [
    "setuptools>=58.0.1",
    "wheel>=0.37.0",
    "toml>=0.10.2",
    "gitpython>=2.1.0",
    'django',
]
build-backend = "setuptools.build_meta"


[project]
name = "rockin"
version = "0.1.0"
description = "A simple data entry webapp to enter rock sample data from wells digged with TU Delft"

dependencies = [
    'mysqlclient',
    'django-widget-tweaks',
    'pydantic',
    'python-dotenv',
]


[project.optional-dependencies]
dev = [
    'coverage',
    'pytest',
    'pytest-cov',
    'pylint',
    'pytest-django',
    'factory-boy',
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "rockin.settings"
